Reset layout/default and cleared index page
Default layout had been updated in zurb-f6-boilerplate so had to reflect that in the project

Removed the og-entries that were previously there by default. Later on they will be added but for now cramped things.

Updated jquery version

Updated .htaccess to remove the x-ua entry since that is now addressed within the actual layout html

Merge and close